Implemented the PR-comment fixes
